improvement(#6): move user dtos to service instead of domain
